Move Serializable to constructs.py
